% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEMevaluator.R
\name{eval.FEM}
\alias{eval.FEM}
\title{Evaluate a FEM object at a set of point locations}
\usage{
eval.FEM(FEM, locations, CPP_CODE = TRUE)
}
\arguments{
\item{FEM}{A \code{FEM} object to be evaluated.}

\item{locations}{A 2-colums(in case of planar mesh) or 3-columns(in case of 2D manifold in a 3D space or a 3D volume) matrix with the spatial locations where the FEM object should be evaluated.}
}
\value{
A matrix of numeric evaluations of the \code{FEM} object. Each row indicates the location where the evaluation has been taken, the column indicates the 
function evaluated.
}
\description{
It evaluates a FEM object the specified set of locations.
}
\references{
Devillers, O. et al. 2001. Walking in a Triangulation, Proceedings of the Seventeenth Annual Symposium on Computational Geometry
}
